feat: 이벤트 포워더 #57

+++ src/main/java/sogang/capstone/editking/infrastructure/event/EventReaderImpl.java
@@ -15,8 +15,8 @@ public class EventReaderImpl implements EventReader {
     private final EventEntryRepository eventEntryRepository;
 
     @Override
-    public List<EventEntry> getEventList(Integer offset, Integer limit) {
-        Pageable limitPageable = PageRequest.of(offset, limit);
+    public List<EventEntry> getEventList(Long offset, Integer limit) {
+        Pageable limitPageable = PageRequest.of(Math.toIntExact(offset), limit);
         return eventEntryRepository.findEventEntryList(limitPageable);
     }
 }
